requirements.txt
numpy
pandas
PyYAML
requests
SQLAlchemy
psycopg2-binary
kafka-python
pyspark
scikit-learn